requirements.txt
alembic==1.7.5
aniso8601==9.0.1
app==0.0.1
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
visitor==0.1.3
Werkzeug==2.0.2
